Add 'imenus-position' structure
There always was a problem: some evil modes (for example,
'markdown-mode') use simple numbers as point positions (instead of
markers) for imenu index, so it was impossible to define from which
buffer a position comes.

This commit introduces a solution: during converting an imenu index to
imenus index, along with renaming, an imenu position is also transformed
into imenus position (a structure which consists of a buffer object and
the imenu position).  Thus all items in 'imenus-index' "knows" to what
buffer they belong.

* imenus.el (imenus-position): New structure.
(imenus-rename-item): Replace with...
(imenus-imenu-item-to-imenus-item): ... this.
(imenus-imenu-index-to-imenus-index, imenus-goto, imenus-buffers):
Adjust accordingly.